refactor browse components into seperate folders
