+++ packages/web/src/app/(app)/browse/[...path]/components/codePreviewPanel/rangeHighlightingExtension.ts
@@ -2,7 +2,7 @@
 
 import { StateField, Range } from "@codemirror/state";
 import { Decoration, DecorationSet, EditorView } from "@codemirror/view";
-import { BrowseHighlightRange } from "../../hooks/utils";
+import { BrowseHighlightRange } from "@/app/(app)/browse/hooks/utils";
 
 const markDecoration = Decoration.mark({
     class: "searchMatch-selected",
